+++ spec/rubocop/cop/layout/rescue_ensure_alignment_spec.rb
@@ -354,6 +354,16 @@ RSpec.describe RuboCop::Cop::Layout::RescueEnsureAlignment, :config do
     RUBY
   end
 
+  it 'accepts correctly aligned rescue in assigned begin-end block' do
+    expect_no_offenses(<<-RUBY)
+      foo = begin
+              bar
+            rescue BazError
+              qux
+            end
+    RUBY
+  end
+
   context '>= Ruby 2.5', :ruby25 do
     it 'accepts aligned rescue in do-end block' do
       expect_no_offenses(<<-RUBY.strip_indent)